# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parking-lot": {
            "get": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "admin get all parking lot",
                "operationId": "index",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/{idParking}": {
            "get": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "detail parking lot with id",
                "operationId": "showParkingLot",
                "parameters": [
                    {
                        "name": "idParking",
                        "in": "path",
                        "description": "ID of the parking lot to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/{id}/info/{userId}": {
            "get": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "detail parking lot with id",
                "operationId": "getInfoParkingLot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking lot to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "IDuser to see status wishlist",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/{id}/info/comment": {
            "get": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "detail comment of parking lot",
                "operationId": "showCommentOfParking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the parking lot to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/location": {
            "get": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "show location near user ",
                "operationId": "showParkingLotNearLocation",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 16.060832
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 108.24149
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/create": {
            "post": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "Create a new parking lot",
                "description": "Create a new parking lot with the specified details",
                "operationId": "createParkingLot",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "userId": {
                                        "type": "integer",
                                        "example": 1000000
                                    },
                                    "images[]": {
                                        "description": "Array of images",
                                        "type": "array",
                                        "items": {
                                            "type": "file"
                                        }
                                    },
                                    "openTime": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "20:08"
                                    },
                                    "endTime": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "21:08"
                                    },
                                    "nameParkingLot": {
                                        "type": "string",
                                        "example": "Parking Lot Cong"
                                    },
                                    "address_latitude": {
                                        "type": "string",
                                        "example": "16.060832"
                                    },
                                    "address_longitude": {
                                        "type": "string",
                                        "example": "108.241491"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "101B Le Huu Tra"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "gia ra dat an ninh coa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking lot created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/update/{idParkingLot}": {
            "post": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "update  parking lot",
                "description": "update parking lot with the specified details",
                "operationId": "updateParkingLot",
                "parameters": [
                    {
                        "name": "idParkingLot",
                        "in": "path",
                        "description": "ID of the parking lot to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000002
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "images[]": {
                                        "description": "Array of images",
                                        "type": "array",
                                        "items": {
                                            "type": "file"
                                        }
                                    },
                                    "openTime": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "20:08"
                                    },
                                    "endTime": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "21:08"
                                    },
                                    "nameParkingLot": {
                                        "type": "string",
                                        "example": "Parking Lot Cong"
                                    },
                                    "address_latitude": {
                                        "type": "string",
                                        "example": "16.060832"
                                    },
                                    "address_longitude": {
                                        "type": "string",
                                        "example": "108.241491"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "101B Le Huu Tra"
                                    },
                                    "desc": {
                                        "type": "string",
                                        "example": "gia ra dat an ninh coa"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking lot created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/parking-lot/delete/{idParkingLot}": {
            "delete": {
                "tags": [
                    "Parking Lot"
                ],
                "summary": "Delete parking lot",
                "operationId": "deleteParkingLot",
                "parameters": [
                    {
                        "name": "idParkingLot",
                        "in": "path",
                        "description": "ID of the parking lot to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1000000
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}